Move EmptyPackages compose to a separate file
