+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/packages/ui/WooShippingCarrierPackageScreen.kt
@@ -2,8 +2,6 @@ package com.woocommerce.android.ui.orders.wooshippinglabels.packages.ui
 
 import android.content.res.Configuration
 import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -40,11 +38,9 @@ import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.woocommerce.android.R
-import com.woocommerce.android.ui.compose.component.WCColoredButton
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
 import com.woocommerce.android.ui.orders.wooshippinglabels.packages.WooShippingLabelPackageCreationViewModel
 import com.woocommerce.android.ui.orders.wooshippinglabels.packages.WooShippingLabelPackageCreationViewModel.PageType
@@ -122,37 +118,6 @@ fun WooShippingCarrierPackageScreen(
             onClick = onAddPackageClick
         ) {
             Text(stringResource(id = R.string.woo_shipping_labels_package_creation_add_package))
-        }
-    }
-}
-
-@Composable
-fun EmptyPackages(
-    modifier: Modifier = Modifier,
-    @DrawableRes image: Int,
-    @StringRes message: Int,
-    onButtonClick: () -> Unit
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colors.surface)
-            .verticalScroll(rememberScrollState())
-            .padding(vertical = 16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Image(painter = painterResource(id = image), contentDescription = null)
-        Spacer(modifier = Modifier.height(14.dp))
-        Text(
-            stringResource(message),
-            style = MaterialTheme.typography.subtitle1,
-            textAlign = TextAlign.Center,
-            color = MaterialTheme.colors.onSurface,
-        )
-        Spacer(modifier = Modifier.height(37.dp))
-        WCColoredButton(onClick = onButtonClick) {
-            Text(stringResource(id = R.string.woo_shipping_labels_package_creation_empty_button))
         }
     }
 }
